fastapi/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
passlib[bcrypt]
bcrypt==4.0.1
python-jose[cryptography]
python-multipart
email-validator